refactor: update migration configuration and remove unused files
