--- src/db/data-source-cli.ts
@@ -1,17 +1,0 @@
-import { DataSource, DataSourceOptions } from 'typeorm';
-import 'dotenv/config';
-
-const dataSourceOptions: DataSourceOptions = {
-  type: 'postgres',
-  host: process.env.DB_HOST,
-  port: Number(process.env.DB_PORT),
-  username: process.env.DB_USERNAME,
-  password: process.env.DB_PASSWORD,
-  database: process.env.DB_NAME,
-  entities: [__dirname + '/../**/entities/*{.js,.ts}'],
-  migrations: [__dirname + '/migrations/*{.js,.ts}'],
-};
-
-const dataSource = new DataSource(dataSourceOptions);
-
-export default dataSource;
